Fix functions with return types not getting tokenized correctly
A function with only a return type but no storage modifier would not be
tokenized as a function (such as `bool:hello()`).  The fix for this was
to move the `storage` include below the function regex.

In addition, `void` is not a valid storage type for variables, so remove
it from the list.